+++ src/main/java/com/slcube/shop/business/member/service/MemberService.java
@@ -1,8 +1,6 @@
 package com.slcube.shop.business.member.service;
 
 import com.slcube.shop.business.member.dto.MemberChangePasswordRequestDto;
-import com.slcube.shop.business.member.dto.MemberLoginDto;
-import com.slcube.shop.business.member.dto.MemberResponseDto;
 import com.slcube.shop.business.member.dto.MemberSignUpRequestDto;
 import org.springframework.stereotype.Service;
 
@@ -10,7 +8,6 @@ import org.springframework.stereotype.Service;
 public interface MemberService {
 
     public Long signUp(MemberSignUpRequestDto requestDto);
-    public MemberResponseDto login(MemberLoginDto requestDto);
 
     public Long changePassword(MemberChangePasswordRequestDto requestDto);
 }
